# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Game Library API",
    "description": "CRUD API for games and game developers with GitHub OAuth authentication.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://cse-341-project2-aa2c.onrender.com",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/auth/success": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/success": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Login through /auth/github to create an authenticated session."
      }
    }
  }
}